Update `vFixedSeeds` with new, reliable nodes
Update `vFixedSeeds` with new nodes that are known and reliable. There may be additional updates to this list in the future. For now, 4 reliable nodes is sufficient for bootstrapping new nodes.

Also clean out the old, dead DNS seeder nodes and add the new DNS endpoint for `seed.lotusia.org`. This DNS endpoint is not active due to issues with the `lotus-seeder` executable and will need to be addressed in the future.